requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
astroid=2.5.8=py38haa95532_0
black=22.6.0=pyhd8ed1ab_0
ca-certificates=2022.07.19=haa95532_0
certifi=2022.6.15=py38haa95532_0
click=8.0.4=py38haa95532_0
colorama=0.4.5=py38haa95532_0
dataclasses=0.8=pyh6d0b6a4_7
isort=5.9.3=pyhd3eb1b0_0
lazy-object-proxy=1.6.0=py38h2bbff1b_0
mccabe=0.6.1=py38_1
mypy=0.961=py38h294d835_0
mypy_extensions=0.4.3=py38haa95532_1
neo4j-python-driver=4.4.5=pyhd8ed1ab_0
openssl=1.1.1q=h2bbff1b_0
pathspec=0.9.0=pyhd8ed1ab_0
pip=22.1.2=py38haa95532_0
platformdirs=2.4.0=pyhd3eb1b0_0
psutil=5.9.0=py38h2bbff1b_0
pylint=2.7.2=py38haa244fe_0
python=3.8.13=h6244533_0
python-dotenv=0.20.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2022.1=py38haa95532_0
setuptools=61.2.0=py38haa95532_0
sqlite=3.38.5=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py38haa95532_0
typed-ast=1.4.3=py38h2bbff1b_1
typing_extensions=4.1.1=pyh06a4308_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
wrapt=1.12.1=py38he774522_1
